Consolidate the middleware files
